Projects/team/Software Engineering/drp/src/db/requirements.txt
mysql-connector-python
python-dotenv
pandas
bcrypt